refactor(napi/oxlint): split into multiple files (#13662)
Pure refactor. In `napi/oxlint2`, split up code into more files, and move all code related to JS plugins (almost all of it) into `plugins` directory.

None of the code is altered, it's just moved around. Only code change is adding a few more comments and 1 type annotation.

This prepares the way for:

1. Lazy-loading the JS plugin code (next PR).
2. Adding a 2nd version of visitor which doesn't use lazy deserialization.
