+++ napi/oxlint2/src-js/plugins/index.ts
@@ -1,0 +1,4 @@
+import { lintFile } from './lint.js';
+import { loadPlugin } from './load.js';
+
+export { lintFile, loadPlugin };
