+++ napi/oxlint2/src-js/plugins/types.ts
