+++ napi/oxlint2/src-js/plugins/utils.ts
