+++ napi/oxlint2/test/compile-visitor.test.ts
@@ -7,9 +7,9 @@ import {
   compiledVisitor,
   finalizeCompiledVisitor,
   initCompiledVisitor,
-} from '../src-js/visitor.js';
+} from '../src-js/plugins/visitor.js';
 
-import type { EnterExit, VisitFn } from '../src-js/types.ts';
+import type { EnterExit, VisitFn } from '../src-js/plugins/types.ts';
 
 const PROGRAM_TYPE_ID = NODE_TYPE_IDS_MAP.get('Program'),
   EMPTY_STMT_TYPE_ID = NODE_TYPE_IDS_MAP.get('EmptyStatement');
